Move extract-over-extend tests into UltimateNormalFormTest
Uses runUnfTest (UnfTransformer directly) instead of
SimplificationTest.runSimplificationTest, matching the workflow already
used for the other UnfTransformer-level tests in this class - a more
direct path to BitvectorUtils than going through the full quantifier
elimination pipeline.
